@@ -1,21 +1,15 @@
-import Vue from 'vue';
-import Vuex from 'vuex';
-
-import Application from './Application.vue';
+import ApplicationComponent from './Application.vue';
 import AppStore from './store/AppStore';
-import Vue2TouchEvents from 'vue2-touch-events';
- 
+import { createStore } from 'vuex';
+import { createApp } from 'vue' ;
 
-Vue.use(Vue2TouchEvents);
-Vue.use(Vuex);
-window.$store = new Vuex.Store({
+const store = createStore({
   modules: {
     App: AppStore,
   },
 });
-Vue.config.productionTip = false;
 
-new Vue({
-  store: window.$store,
-  render: (h) => h(Application),
-}).$mount('#app');
+const app = createApp(ApplicationComponent);
+
+app.use(store);
+app.mount('#app');
